Move page edit view to wagtail.admin.views.pages.edit
